@@ -22,10 +22,11 @@
     "webpack-cli": "^4.9.1"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.80.6",
     "css-loader": "^7.1.2",
     "postcss-loader": "^8.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "style-loader": "^4.0.0"
   }
 }
