@@ -1,5 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import { createRoot } from "react-dom/client";
+import App from "./popup/App";
 import "./index.css";
-ReactDOM.render(<App />, document.getElementById('root-ext'));
+import React from "react";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+const container = document.getElementById("root");
+const root = createRoot(container!);
+const queryClient = new QueryClient();
+
+root.render(
+  <React.StrictMode>
+    <QueryClientProvider client={queryClient}>
+      <App />
+    </QueryClientProvider>
+  </React.StrictMode>
+);
